% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_S3.R
\name{AIC.CopulaCenR}
\alias{AIC.CopulaCenR}
\title{the AIC of a CopulaCenR object}
\usage{
\method{AIC}{CopulaCenR}(object, ..., k = 2)
}
\arguments{
\item{object}{a CopulaCenR object}

\item{...}{further arguments}

\item{k}{numeric, with k = 2 for AIC}
}
\description{
the AIC of a CopulaCenR object
}
