% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_S3.R
\name{BIC.CopulaCenR}
\alias{BIC.CopulaCenR}
\title{the BIC of a CopulaCenR object}
\usage{
\method{BIC}{CopulaCenR}(object, ...)
}
\arguments{
\item{object}{a CopulaCenR object}

\item{...}{further arguments}
}
\description{
the BIC of a CopulaCenR object
}
