% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_S3.R
\name{coef.CopulaCenR}
\alias{coef.CopulaCenR}
\title{the coefficient estimates of an CopulaCenR object}
\usage{
\method{coef}{CopulaCenR}(object, ...)
}
\arguments{
\item{object}{a CopulaCenR object}

\item{...}{further arguments}
}
\description{
the coefficient estimates of an CopulaCenR object
}
