\name{data_sim_multi_rec}
\alias{data_sim_multi_rec}
\docType{data}
\title{
  data_sim_multi_rec
}
\description{
  A simulated real dataset of multivariate recurrent events data with 300 subjects.
}
\usage{data("data_sim_multi_rec")}
\format{
  A list with 300 subjects on the following 2 components.
  \describe{
    \item{\code{gaptimes}}{a data frame with 300 subjects in row and 4 gap times in column}
    \item{\code{status}}{a data frame with 300 subjects in row and 4 statuses in column}
  }
}

\keyword{datasets}
