% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_S3.R
\name{logLik.CopulaCenR}
\alias{logLik.CopulaCenR}
\title{the log-likelihood of an CopulaCenR object}
\usage{
\method{logLik}{CopulaCenR}(object, ...)
}
\arguments{
\item{object}{a CopulaCenR object}

\item{...}{further arguments}
}
\description{
the log-likelihood of an CopulaCenR object
}
