% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_S3.R
\name{print.CopulaCenR}
\alias{print.CopulaCenR}
\title{Printing outputs of a CopulaCenR object}
\usage{
\method{print}{CopulaCenR}(x, ...)
}
\arguments{
\item{x}{a CopulaCenR object}

\item{...}{further arguments}
}
\description{
Printing outputs of a CopulaCenR object
}
