% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_S3.R
\name{print.summary.CopulaCenR}
\alias{print.summary.CopulaCenR}
\title{Print the summary of an CopulaCenR object}
\usage{
\method{print}{summary.CopulaCenR}(x, ...)
}
\arguments{
\item{x}{a CopulaCenR object}

\item{...}{further arguments}
}
\description{
Print the summary of an CopulaCenR object
}
