% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_S3.R
\name{summary.CopulaCenR}
\alias{summary.CopulaCenR}
\title{Summarizing outputs of a CopulaCenR object}
\usage{
\method{summary}{CopulaCenR}(object, ...)
}
\arguments{
\item{object}{a CopulaCenR object}

\item{...}{further arguments}
}
\description{
Summarizing outputs of a CopulaCenR object
}
