% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_copula.R
\name{tau_copula}
\alias{tau_copula}
\title{Calculate Kendall's tau}
\usage{
tau_copula(eta, copula)
}
\arguments{
\item{eta}{copula parameter(s); if Coupla2, input as a vector of (alpha, kappa)}

\item{copula}{specify the type of copula model}
}
\value{
Kendall's tau
}
\description{
This function (tau_copula) is to obtain Kendall's tau from copula parameter(s)
}
